+++ junos/resource_system_root_authentication.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"golang.org/x/crypto/ssh"
 )
 
@@ -27,8 +28,17 @@ func resourceSystemRootAuthentication() *schema.Resource {
 		},
 		Schema: map[string]*schema.Schema{
 			"encrypted_password": {
-				Type:     schema.TypeString,
-				Required: true,
+				Type:         schema.TypeString,
+				Optional:     true,
+				ValidateFunc: validation.StringIsNotEmpty,
+				ExactlyOneOf: []string{"encrypted_password", "plain_text_password"},
+			},
+			"plain_text_password": {
+				Type:         schema.TypeString,
+				Optional:     true,
+				Sensitive:    true,
+				ValidateFunc: validation.StringIsNotEmpty,
+				ExactlyOneOf: []string{"encrypted_password", "plain_text_password"},
 			},
 			"no_public_keys": {
 				Type:     schema.TypeBool,
@@ -51,6 +61,12 @@ func resourceSystemRootAuthenticationCreate(ctx context.Context, d *schema.Resou
 ) diag.Diagnostics {
 	sess := m.(*Session)
 	if sess.junosFakeCreateSetFile != "" {
+		// To be able detect a plain text password not accepted by system
+		if d.Get("plain_text_password").(string) != "" {
+			if err := delSystemRootAuthenticationPassword(m, nil); err != nil {
+				return diag.FromErr(err)
+			}
+		}
 		if err := setSystemRootAuthentication(d, m, nil); err != nil {
 			return diag.FromErr(err)
 		}
@@ -65,6 +81,14 @@ func resourceSystemRootAuthenticationCreate(ctx context.Context, d *schema.Resou
 	defer sess.closeSession(jnprSess)
 	sess.configLock(jnprSess)
 	var diagWarns diag.Diagnostics
+	// To be able detect a plain text password not accepted by system
+	if d.Get("plain_text_password").(string) != "" {
+		if err := delSystemRootAuthenticationPassword(m, jnprSess); err != nil {
+			appendDiagWarns(&diagWarns, sess.configClear(jnprSess))
+
+			return append(diagWarns, diag.FromErr(err)...)
+		}
+	}
 	if err := setSystemRootAuthentication(d, m, jnprSess); err != nil {
 		appendDiagWarns(&diagWarns, sess.configClear(jnprSess))
 
@@ -180,7 +204,11 @@ func setSystemRootAuthentication(d *schema.ResourceData, m interface{}, jnprSess
 	configSet := make([]string, 0)
 	setPrefix := "set system root-authentication "
 
-	configSet = append(configSet, setPrefix+"encrypted-password \""+d.Get("encrypted_password").(string)+"\"")
+	if v := d.Get("plain_text_password").(string); v != "" {
+		configSet = append(configSet, setPrefix+"plain-text-password-value \""+v+"\"")
+	} else {
+		configSet = append(configSet, setPrefix+"encrypted-password \""+d.Get("encrypted_password").(string)+"\"")
+	}
 	if d.Get("no_public_keys").(bool) {
 		configSet = append(configSet, setPrefix+"no-public-keys")
 	}
@@ -253,9 +281,19 @@ func delSystemRootAuthentication(m interface{}, jnprSess *NetconfObject) error {
 	return sess.configSet(configSet, jnprSess)
 }
 
+func delSystemRootAuthenticationPassword(m interface{}, jnprSess *NetconfObject) error {
+	sess := m.(*Session)
+	configSet := make([]string, 0, 1)
+	configSet = append(configSet, "delete system root-authentication encrypted-password")
+
+	return sess.configSet(configSet, jnprSess)
+}
+
 func fillSystemRootAuthenticationData(d *schema.ResourceData, systemRootAuthOptions systemRootAuthOptions) {
-	if tfErr := d.Set("encrypted_password", systemRootAuthOptions.encryptedPassword); tfErr != nil {
-		panic(tfErr)
+	if d.Get("plain_text_password").(string) == "" {
+		if tfErr := d.Set("encrypted_password", systemRootAuthOptions.encryptedPassword); tfErr != nil {
+			panic(tfErr)
+		}
 	}
 	if tfErr := d.Set("no_public_keys", systemRootAuthOptions.noPublicKeys); tfErr != nil {
 		panic(tfErr)